refactor(metrics): observed metrics move into their own package
A metric measures and a grader judges, and the measuring half now has
one home: @egma/metrics holds the versioned measure catalog and the one
shared measure module that computes it, for a simulation and a
production trace alike. The catalog implementation leaves
@egma/simulation-contract, which returns to exactly what its name says;
the arithmetic leaves @egma/db, where it lived by dependency convenience
and did no data access. The module's span slice is declared structurally
in-package, so the data-access module depends on metrics and never the
reverse, and everySpanIn walks any span tree generically. The
single-implementation guard and the simulator drift tests move with the
module they pin.
