+++ apps/grader/package.json
@@ -11,13 +11,14 @@
   "dependencies": {
     "@egma/db": "workspace:*",
     "@egma/ids": "workspace:*",
+    "@egma/metrics": "workspace:*",
     "@egma/provider-credentials": "workspace:*",
     "@egma/simulation-contract": "workspace:*",
     "@opentelemetry/api": "^1.9.0",
     "@opentelemetry/instrumentation": "^0.221.0",
     "@opentelemetry/instrumentation-http": "^0.221.0",
-    "@opentelemetry/instrumentation-pino": "^0.67.0",
     "@opentelemetry/instrumentation-pg": "^0.73.0",
+    "@opentelemetry/instrumentation-pino": "^0.67.0",
     "@opentelemetry/instrumentation-undici": "^0.31.0",
     "@opentelemetry/sdk-node": "^0.221.0",
     "pino": "^10.3.1",
