+++ package.json
@@ -11,9 +11,9 @@
     "skills:sync": "node tools/project-agent-skills.mjs --write",
     "skills:check": "node tools/project-agent-skills.mjs --check",
     "lint": "node packages/lint/src/cli.ts",
-    "build": "pnpm skills:check && pnpm --filter @egma/platform-api generate:check && pnpm lint && tsc -b packages/ids packages/platform-api packages/provider-credentials packages/retell packages/db packages/lint apps/api apps/grader apps/cli",
+    "build": "pnpm skills:check && pnpm --filter @egma/platform-api generate:check && pnpm lint && tsc -b packages/ids packages/metrics packages/platform-api packages/provider-credentials packages/retell packages/db packages/lint apps/api apps/grader apps/cli",
     "typecheck": "pnpm build && tsc -p tsconfig.tests.json && pnpm --filter @egma/web typecheck",
-    "clean": "tsc -b --clean packages/ids packages/platform-api packages/provider-credentials packages/retell packages/db packages/lint apps/api apps/grader apps/cli",
+    "clean": "tsc -b --clean packages/ids packages/metrics packages/platform-api packages/provider-credentials packages/retell packages/db packages/lint apps/api apps/grader apps/cli",
     "db:up": "node packages/db/test/support/start-stores.ts && node packages/db/test/support/sweep-stale-databases.ts",
     "db:down": "node packages/db/test/support/stop-stores.ts",
     "db:generate": "pnpm --filter @egma/db generate",
