+++ packages/db/test/graders.test.ts
@@ -2,7 +2,8 @@ import { isId, newId } from "@egma/ids";
 import {
   CATALOGED_MEASURES,
   SPAN_DERIVED_MEASURES,
-} from "@egma/simulation-contract";
+} from "@egma/metrics";
+
 import { afterAll, beforeAll, describe, expect, it } from "vitest";
 
 import {
