+++ packages/db/tsconfig.json
@@ -5,6 +5,18 @@
     "rootDir": "./src",
     "tsBuildInfoFile": "./dist/.tsbuildinfo"
   },
-  "include": ["src/**/*.ts"],
-  "references": [{ "path": "../ids" }, { "path": "../simulation-contract" }]
+  "include": [
+    "src/**/*.ts"
+  ],
+  "references": [
+    {
+      "path": "../ids"
+    },
+    {
+      "path": "../metrics"
+    },
+    {
+      "path": "../simulation-contract"
+    }
+  ]
 }
