+++ packages/metrics/tsconfig.json
@@ -5,15 +5,5 @@
     "rootDir": "./src",
     "tsBuildInfoFile": "./dist/.tsbuildinfo"
   },
-  "include": [
-    "src/**/*.ts"
-  ],
-  "references": [
-    {
-      "path": "../ids"
-    },
-    {
-      "path": "../metrics"
-    }
-  ]
+  "include": ["src/**/*.ts"]
 }
