+++ tsconfig.tests.json
@@ -8,6 +8,7 @@
     "baseUrl": ".",
     "paths": {
       "@egma/ids": ["./packages/ids/src/index.ts"],
+      "@egma/metrics": ["./packages/metrics/src/index.ts"],
       "@egma/provider-credentials": [
         "./packages/provider-credentials/src/index.ts"
       ],
